feat: local test용 컴포넌트 추가
