+++ src/components/common/Modal/FilterModal/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './FilterModal';
