+++ src/components/common/Modal/SortModal.tsx
@@ -1,5 +1,7 @@
 import React from 'react'
 
+
+
 const SortModal = () => {
   return (
     <div>
